Additional Ammunition
- Separated magazines into files for ease of access.
- Added 5.8x42 DBP88 ammo. (30 & 100Rnd variants)
- Added 2Rnd Magnum shells.
