+++ addons/ammunition/556x45.hpp
@@ -1,0 +1,339 @@
+// AK Mags
+class CLASS(30Rnd_556x45_M855A1_AK): CUP_30Rnd_556x45_AK {
+    author = "TyroneMF";
+    scope = 2;
+    lastRoundsTracer = 4;
+    displayName = CSTRING(30Rnd_556x45_M855A1_AK_Name);
+    ammo = QCLASS(556x45_EPR_M855A1);
+    mass = 10;
+};
+
+class CLASS(30Rnd_556x45_M855A1_AK_Red): CLASS(30Rnd_556x45_M855A1_AK) {
+    displayName = CSTRING(30Rnd_556x45_M855A1_AK_Red_Name);
+    tracersEvery = 1;
+    ammo = QCLASS(556x45_EPR_M855A1_Red);
+};
+
+class CLASS(30Rnd_556x45_M995_AK): CLASS(30Rnd_556x45_M855A1_AK) {
+    displayName = CSTRING(30Rnd_556x45_M995_AK_Name);
+    ammo = "ACE_556x45_Ball_M995_AP";
+};
+
+// AUG Mags
+class CLASS(30Rnd_556x45_M855A1_AUG): CUP_30Rnd_556x45_AUG {
+    author = "TyroneMF";
+    scope = 2;
+    lastRoundsTracer = 4;
+    displayName = CSTRING(30Rnd_556x45_M855A1_AUG_Name);
+    ammo = QCLASS(556x45_EPR_M855A1);
+    mass = 10;
+};
+
+class CLASS(30Rnd_556x45_M855A1_AUG_Red): CLASS(30Rnd_556x45_M855A1_AUG) {
+    displayName = CSTRING(30Rnd_556x45_M855A1_AUG_Red_Name);
+    tracersEvery = 1;
+    ammo = QCLASS(556x45_EPR_M855A1_Red);
+};
+
+class CLASS(30Rnd_556x45_M995_AUG): CLASS(30Rnd_556x45_M855A1_AUG) {
+    displayName = CSTRING(30Rnd_556x45_M995_AUG_Name);
+    ammo = "ACE_556x45_Ball_M995_AP";
+};
+
+// G36 Mags
+class CLASS(30Rnd_556x45_M855A1_G36): CUP_30Rnd_556x45_G36 {
+    author = "TyroneMF";
+    scope = 2;
+    lastRoundsTracer = 4;
+    displayName = CSTRING(30Rnd_556x45_M855A1_G36_Name);
+    ammo = QCLASS(556x45_EPR_M855A1);
+    mass = 10;
+};
+
+class CLASS(30Rnd_556x45_M855A1_G36_Red): CLASS(30Rnd_556x45_M855A1_G36) {
+    displayName = CSTRING(30Rnd_556x45_M855A1_G36_Red_Name);
+    ammo = QCLASS(556x45_EPR_M855A1_Red);
+    tracersEvery = 1;
+};
+
+class CLASS(30Rnd_556x45_M995_G36): CLASS(30Rnd_556x45_M855A1_G36) {
+    displayName = CSTRING(30Rnd_556x45_M995_G36_Name);
+    ammo = "ACE_556x45_Ball_M995_AP";
+};
+
+// SIG Magazines
+class CLASS(30Rnd_556x45_M855A1_SG550): hlc_30Rnd_556x45_EPR_sg550 {
+    author = "TyroneMF";
+    scope = 2;
+    lastRoundsTracer = 4;
+    displayName = CSTRING(30Rnd_556x45_M855A1_SG550_Name);
+    ammo = QCLASS(556x45_EPR_M855A1);
+    mass = 10;
+};
+
+class CLASS(30Rnd_556x45_M855A1_SG550_Red): CLASS(30Rnd_556x45_M855A1_SG550) {
+    displayName = CSTRING(30Rnd_556x45_M855A1_SG550_Red_Name);
+    tracersEvery = 1;
+    ammo = QCLASS(556x45_EPR_M855A1_Red);
+};
+
+class CLASS(30Rnd_556x45_M995_SG550): CLASS(30Rnd_556x45_M855A1_SG550) {
+    displayName = CSTRING(30Rnd_556x45_M995_SG550_Name);
+    ammo = "ACE_556x45_Ball_M995_AP";
+};
+
+// PMAG
+class CLASS(30Rnd_556x45_M855_PMAG): 30Rnd_556x45_Stanag_green {
+    author = "TyroneMF";
+    scope = 2;
+    lastRoundsTracer = 4;
+    picture = "\CUP\Weapons\CUP_Weapons_Ammunition\data\ui\m_pmag_ca.paa";
+    displayName = CSTRING(30Rnd_556x45_M855_PMAG_Name);
+    mass = 10;
+    model = "\CUP\Weapons\CUP_Weapons_Ammunition\magazines\CUP_mag_30rnd_pmag_qp.p3d";
+    modelSpecial = "\CUP\Weapons\CUP_Weapons_Ammunition\magazines_proxy\CUP_mag_30rnd_pmag_qp.p3d";
+    modelSpecialIsProxy = 1;
+    hiddenSelections[] = {"Camo1"};
+    hiddenSelectionsTextures[] = {"\CUP\Weapons\CUP_Weapons_Ammunition\magazines\data\pmag_black_co.paa"};
+};
+
+class CLASS(30Rnd_556x45_M855_PMAG_Tan): CLASS(30Rnd_556x45_M855_PMAG) {
+    picture = "\CUP\Weapons\CUP_Weapons_Ammunition\data\ui\m_pmag_coyote_ca.paa";
+    displayName = CSTRING(30Rnd_556x45_M855_PMAG_Tan_Name);
+    hiddenSelections[] = {"Camo1"};
+    hiddenSelectionsTextures[] = {"\CUP\Weapons\CUP_Weapons_Ammunition\magazines\data\pmag_coyote_co.paa"};
+};
+
+class CLASS(30Rnd_556x45_M855A1_PMAG): 30Rnd_556x45_Stanag_green {
+    author = "TyroneMF";
+    scope = 2;
+    lastRoundsTracer = 4;
+    picture = "\CUP\Weapons\CUP_Weapons_Ammunition\data\ui\m_pmag_ca.paa";
+    displayName = CSTRING(30Rnd_556x45_M855A1_PMAG_Name);
+    ammo = QCLASS(556x45_EPR_M855A1);
+    mass = 10;
+    model = "\CUP\Weapons\CUP_Weapons_Ammunition\magazines\CUP_mag_30rnd_pmag_qp.p3d";
+    modelSpecial = "\CUP\Weapons\CUP_Weapons_Ammunition\magazines_proxy\CUP_mag_30rnd_pmag_qp.p3d";
+    modelSpecialIsProxy = 1;
+    hiddenSelections[] = {"Camo1"};
+    hiddenSelectionsTextures[] = {"\CUP\Weapons\CUP_Weapons_Ammunition\magazines\data\pmag_black_co.paa"};
+};
+
+class CLASS(30Rnd_556x45_M855A1_PMAG_Tan): CLASS(30Rnd_556x45_M855A1_PMAG) {
+    picture = "\CUP\Weapons\CUP_Weapons_Ammunition\data\ui\m_pmag_coyote_ca.paa";
+    displayName = CSTRING(30Rnd_556x45_M855A1_PMAG_Tan_Name);
+    hiddenSelections[] = {"Camo1"};
+    hiddenSelectionsTextures[] = {"\CUP\Weapons\CUP_Weapons_Ammunition\magazines\data\pmag_coyote_co.paa"};
+};
+
+class CLASS(30Rnd_556x45_M995_PMAG): ACE_30Rnd_556x45_Stanag_M995_AP_mag {
+    author = "TyroneMF";
+    scope = 2;
+    lastRoundsTracer = 2;
+    picture = "\CUP\Weapons\CUP_Weapons_Ammunition\data\ui\m_pmag_ca.paa";
+    displayName = CSTRING(30Rnd_556x45_M995_PMAG_Name);
+    mass = 10;
+    model = "\CUP\Weapons\CUP_Weapons_Ammunition\magazines\CUP_mag_30rnd_pmag_qp.p3d";
+    modelSpecial = "\CUP\Weapons\CUP_Weapons_Ammunition\magazines_proxy\CUP_mag_30rnd_pmag_qp.p3d";
+    modelSpecialIsProxy = 1;
+    hiddenSelections[] = {"Camo1"};
+    hiddenSelectionsTextures[] = {"\CUP\Weapons\CUP_Weapons_Ammunition\magazines\data\pmag_black_co.paa"};
+};
+
+class CLASS(30Rnd_556x45_M995_PMAG_Tan): CLASS(30Rnd_556x45_M995_PMAG) {
+    picture = "\CUP\Weapons\CUP_Weapons_Ammunition\data\ui\m_pmag_coyote_ca.paa";
+    displayName = CSTRING(30Rnd_556x45_M995_PMAG_Tan_Name);
+    hiddenSelections[] = {"Camo1"};
+    hiddenSelectionsTextures[] = {"\CUP\Weapons\CUP_Weapons_Ammunition\magazines\data\pmag_coyote_co.paa"};
+};
+
+class CLASS(30Rnd_556x45_MK262_PMAG): ACE_30Rnd_556x45_Stanag_Mk262_mag {
+    author = "TyroneMF";
+    scope = 2;
+    lastRoundsTracer = 2;
+    picture = "\CUP\Weapons\CUP_Weapons_Ammunition\data\ui\m_pmag_ca.paa";
+    displayName = CSTRING(30Rnd_556x45_MK262_PMAG_Name);
+    mass = 10;
+    model = "\CUP\Weapons\CUP_Weapons_Ammunition\magazines\CUP_mag_30rnd_pmag_qp.p3d";
+    modelSpecial = "\CUP\Weapons\CUP_Weapons_Ammunition\magazines_proxy\CUP_mag_30rnd_pmag_qp.p3d";
+    modelSpecialIsProxy = 1;
+    hiddenSelections[] = {"Camo1"};
+    hiddenSelectionsTextures[] = {"\CUP\Weapons\CUP_Weapons_Ammunition\magazines\data\pmag_black_co.paa"};
+};
+
+class CLASS(30Rnd_556x45_MK262_PMAG_Tan): CLASS(30Rnd_556x45_MK262_PMAG) {
+    picture = "\CUP\Weapons\CUP_Weapons_Ammunition\data\ui\m_pmag_coyote_ca.paa";
+    displayName = CSTRING(30Rnd_556x45_MK262_PMAG_Tan_Name);
+    hiddenSelections[] = {"Camo1"};
+    hiddenSelectionsTextures[] = {"\CUP\Weapons\CUP_Weapons_Ammunition\magazines\data\pmag_coyote_co.paa"};
+};
+
+class CLASS(30Rnd_556x45_MK318_PMAG): ACE_30Rnd_556x45_Stanag_Mk318_mag {
+    author = "TyroneMF";
+    scope = 2;
+    lastRoundsTracer = 2;
+    picture = "\CUP\Weapons\CUP_Weapons_Ammunition\data\ui\m_pmag_ca.paa";
+    displayName = CSTRING(30Rnd_556x45_MK318_PMAG_Name);
+    mass = 10;
+    model = "\CUP\Weapons\CUP_Weapons_Ammunition\magazines\CUP_mag_30rnd_pmag_qp.p3d";
+    modelSpecial = "\CUP\Weapons\CUP_Weapons_Ammunition\magazines_proxy\CUP_mag_30rnd_pmag_qp.p3d";
+    modelSpecialIsProxy = 1;
+    hiddenSelections[] = {"Camo1"};
+    hiddenSelectionsTextures[] = {"\CUP\Weapons\CUP_Weapons_Ammunition\magazines\data\pmag_black_co.paa"};
+};
+
+class CLASS(30Rnd_556x45_MK318_PMAG_Tan): CLASS(30Rnd_556x45_MK318_PMAG) {
+    picture = "\CUP\Weapons\CUP_Weapons_Ammunition\data\ui\m_pmag_coyote_ca.paa";
+    displayName = CSTRING(30Rnd_556x45_MK318_PMAG_Tan_Name);
+    hiddenSelections[] = {"Camo1"};
+    hiddenSelectionsTextures[] = {"\CUP\Weapons\CUP_Weapons_Ammunition\magazines\data\pmag_coyote_co.paa"};
+};
+
+// EMAG
+class CLASS(30Rnd_556x45_M855_EMAG): 30Rnd_556x45_Stanag_green {
+    author = "TyroneMF";
+    scope = 2;
+    lastRoundsTracer = 4;
+    picture = "\CUP\Weapons\CUP_Weapons_Ammunition\data\ui\m_emag_ca.paa";
+    displayName = CSTRING(30Rnd_556x45_M855_EMAG_Name);
+    mass = 10;
+    model = "\CUP\Weapons\CUP_Weapons_Ammunition\magazines\CUP_mag_30Rnd_EMAG.p3d";
+    modelSpecial = "\CUP\Weapons\CUP_Weapons_Ammunition\magazines_proxy\CUP_mag_30Rnd_EMAG.p3d";
+    modelSpecialIsProxy = 1;
+    hiddenSelections[] = {"Camo1"};
+    hiddenSelectionsTextures[] = {"\CUP\Weapons\CUP_Weapons_Ammunition\magazines\data\emag_co.paa"};
+};
+
+class CLASS(30Rnd_556x45_M855_EMAG_Tan): CLASS(30Rnd_556x45_M855_EMAG) {
+    picture = "\CUP\Weapons\CUP_Weapons_Ammunition\data\ui\m_emag_coyote_ca.paa";
+    displayName = CSTRING(30Rnd_556x45_M855_EMAG_Tan_Name);
+    model = "\CUP\Weapons\CUP_Weapons_Ammunition\magazines\CUP_mag_30Rnd_EMAG_coyote.p3d";
+    hiddenSelections[] = {"Camo1"};
+    hiddenSelectionsTextures[] = {"\CUP\Weapons\CUP_Weapons_Ammunition\magazines\data\emag_coyote_co.paa"};
+};
+
+class CLASS(30Rnd_556x45_M855A1_EMAG): 30Rnd_556x45_Stanag_green {
+    author = "TyroneMF";
+    scope = 2;
+    lastRoundsTracer = 4;
+    picture = "\CUP\Weapons\CUP_Weapons_Ammunition\data\ui\m_emag_ca.paa";
+    displayName = CSTRING(30Rnd_556x45_M855A1_EMAG_Name);
+    ammo = QCLASS(556x45_EPR_M855A1);
+    mass = 10;
+    model = "\CUP\Weapons\CUP_Weapons_Ammunition\magazines\CUP_mag_30Rnd_EMAG.p3d";
+    modelSpecial = "\CUP\Weapons\CUP_Weapons_Ammunition\magazines_proxy\CUP_mag_30Rnd_EMAG.p3d";
+    modelSpecialIsProxy = 1;
+    hiddenSelections[] = {"Camo1"};
+    hiddenSelectionsTextures[] = {"\CUP\Weapons\CUP_Weapons_Ammunition\magazines\data\emag_co.paa"};
+};
+
+class CLASS(30Rnd_556x45_M855A1_EMAG_Tan): CLASS(30Rnd_556x45_M855A1_EMAG) {
+    picture = "\CUP\Weapons\CUP_Weapons_Ammunition\data\ui\m_emag_coyote_ca.paa";
+    displayName = CSTRING(30Rnd_556x45_M855A1_EMAG_Tan_Name);
+    model = "\CUP\Weapons\CUP_Weapons_Ammunition\magazines\CUP_mag_30Rnd_EMAG_coyote.p3d";
+    hiddenSelections[] = {"Camo1"};
+    hiddenSelectionsTextures[] = {"\CUP\Weapons\CUP_Weapons_Ammunition\magazines\data\emag_coyote_co.paa"};
+};
+
+class CLASS(30Rnd_556x45_M995_EMAG): ACE_30Rnd_556x45_Stanag_M995_AP_mag {
+    author = "TyroneMF";
+    scope = 2;
+    lastRoundsTracer = 2;
+    picture = "\CUP\Weapons\CUP_Weapons_Ammunition\data\ui\m_emag_ca.paa";
+    displayName = CSTRING(30Rnd_556x45_M995_EMAG_Name);
+    mass = 10;
+    model = "\CUP\Weapons\CUP_Weapons_Ammunition\magazines\CUP_mag_30Rnd_EMAG.p3d";
+    modelSpecial = "\CUP\Weapons\CUP_Weapons_Ammunition\magazines_proxy\CUP_mag_30Rnd_EMAG.p3d";
+    modelSpecialIsProxy = 1;
+    hiddenSelections[] = {"Camo1"};
+    hiddenSelectionsTextures[] = {"\CUP\Weapons\CUP_Weapons_Ammunition\magazines\data\emag_co.paa"};
+};
+
+class CLASS(30Rnd_556x45_M995_EMAG_Tan): CLASS(30Rnd_556x45_M995_EMAG) {
+    picture = "\CUP\Weapons\CUP_Weapons_Ammunition\data\ui\m_emag_coyote_ca.paa";
+    displayName = CSTRING(30Rnd_556x45_M995_EMAG_Tan_Name);
+    model = "\CUP\Weapons\CUP_Weapons_Ammunition\magazines\CUP_mag_30Rnd_EMAG_coyote.p3d";
+    hiddenSelections[] = {"Camo1"};
+    hiddenSelectionsTextures[] = {"\CUP\Weapons\CUP_Weapons_Ammunition\magazines\data\emag_coyote_co.paa"};
+};
+
+class CLASS(30Rnd_556x45_MK262_EMAG): ACE_30Rnd_556x45_Stanag_Mk262_mag {
+    author = "TyroneMF";
+    scope = 2;
+    lastRoundsTracer = 2;
+    picture = "\CUP\Weapons\CUP_Weapons_Ammunition\data\ui\m_emag_ca.paa";
+    displayName = CSTRING(30Rnd_556x45_MK262_EMAG_Name);
+    mass = 10;
+    model = "\CUP\Weapons\CUP_Weapons_Ammunition\magazines\CUP_mag_30Rnd_EMAG.p3d";
+    modelSpecial = "\CUP\Weapons\CUP_Weapons_Ammunition\magazines_proxy\CUP_mag_30Rnd_EMAG.p3d";
+    modelSpecialIsProxy = 1;
+    hiddenSelections[] = {"Camo1"};
+    hiddenSelectionsTextures[] = {"\CUP\Weapons\CUP_Weapons_Ammunition\magazines\data\emag_co.paa"};
+};
+
+class CLASS(30Rnd_556x45_MK262_EMAG_Tan): CLASS(30Rnd_556x45_MK262_EMAG) {
+    picture = "\CUP\Weapons\CUP_Weapons_Ammunition\data\ui\m_emag_coyote_ca.paa";
+    displayName = CSTRING(30Rnd_556x45_MK262_EMAG_Tan_Name);
+    model = "\CUP\Weapons\CUP_Weapons_Ammunition\magazines\CUP_mag_30Rnd_EMAG_coyote.p3d";
+    hiddenSelections[] = {"Camo1"};
+    hiddenSelectionsTextures[] = {"\CUP\Weapons\CUP_Weapons_Ammunition\magazines\data\emag_coyote_co.paa"};
+};
+
+class CLASS(30Rnd_556x45_MK318_EMAG): ACE_30Rnd_556x45_Stanag_Mk318_mag {
+    author = "TyroneMF";
+    scope = 2;
+    lastRoundsTracer = 2;
+    picture = "\CUP\Weapons\CUP_Weapons_Ammunition\data\ui\m_emag_ca.paa";
+    displayName = CSTRING(30Rnd_556x45_MK318_EMAG_Name);
+    mass = 10;
+    model = "\CUP\Weapons\CUP_Weapons_Ammunition\magazines\CUP_mag_30Rnd_EMAG.p3d";
+    modelSpecial = "\CUP\Weapons\CUP_Weapons_Ammunition\magazines_proxy\CUP_mag_30Rnd_EMAG.p3d";
+    modelSpecialIsProxy = 1;
+    hiddenSelections[] = {"Camo1"};
+    hiddenSelectionsTextures[] = {"\CUP\Weapons\CUP_Weapons_Ammunition\magazines\data\emag_co.paa"};
+};
+
+class CLASS(30Rnd_556x45_MK318_EMAG_Tan): CLASS(30Rnd_556x45_MK318_EMAG) {
+    picture = "\CUP\Weapons\CUP_Weapons_Ammunition\data\ui\m_emag_coyote_ca.paa";
+    displayName = CSTRING(30Rnd_556x45_MK318_EMAG_Tan_Name);
+    model = "\CUP\Weapons\CUP_Weapons_Ammunition\magazines\CUP_mag_30Rnd_EMAG_coyote.p3d";
+    hiddenSelections[] = {"Camo1"};
+    hiddenSelectionsTextures[] = {"\CUP\Weapons\CUP_Weapons_Ammunition\magazines\data\emag_coyote_co.paa"};
+};
+
+// 60Rnd 556 Magazines
+class CLASS(60Rnd_556x45_M855A1_SureFire): CUP_60Rnd_556x45_SureFire {
+    author = "TyroneMF";
+    scope = 2;
+    displayName = CSTRING(60Rnd_556x45_M855A1_SureFire_Name);
+    ammo = QCLASS(556x45_EPR_M855A1);
+    tracersEvery = 4;
+};
+
+// 200Rnd 556 Box Magazines
+class CLASS(200Rnd_556x45_M855A1_Box): 200Rnd_556x45_Box_F {
+    author = "TyroneMF";
+    scope = 2;
+    displayName = CSTRING(200Rnd_556x45_M855A1_Box_Name);
+    ammo = QCLASS(556x45_EPR_M855A1);
+    mass = 40;
+    tracersEvery = 4;
+};
+
+class CLASS(200Rnd_556x45_M855A1_Box_Red): CLASS(200Rnd_556x45_M855A1_Box) {
+    displayName = CSTRING(200Rnd_556x45_M855A1_Box_Red_Name);
+    ammo = QCLASS(556x45_EPR_M855A1_Red);
+    tracersEvery = 4;
+};
+
+class CLASS(200Rnd_556x45_M855A1_Box_Tracer): CLASS(200Rnd_556x45_M855A1_Box) {
+    displayName = CSTRING(200Rnd_556x45_M855A1_Box_Tracer_Name);
+    ammo = QCLASS(556x45_EPR_M855A1);
+    tracersEvery = 1;
+};
+
+class CLASS(200Rnd_556x45_M855A1_Box_Tracer_Red): CLASS(200Rnd_556x45_M855A1_Box_Tracer) {
+    displayName = CSTRING(200Rnd_556x45_M855A1_Box_Tracer_Red_Name);
+    ammo = QCLASS(556x45_EPR_M855A1_Red);
+    tracersEvery = 1;
+};
