+++ addons/ammunition/762x51.hpp
@@ -1,0 +1,17 @@
+// 20Rnd 762x51 Magazines
+class CLASS(20Rnd_762x51_FAL_AP_Mag): CUP_20Rnd_762x51_FNFAL_M {
+    author = "TyroneMF";
+    scope = 2;
+    displayName = CSTRING(20Rnd_762x51_FAL_AP_Mag_Name);
+    ammo = "ACE_762x51_Ball_M993_AP";
+    lastRoundsTracer = 2;
+};
+
+// 30Rnd 762x51 Magazines
+class CLASS(30Rnd_762x51_FAL_AP_Mag): CUP_30Rnd_762x51_FNFAL_M {
+    author = "TyroneMF";
+    scope = 2;
+    displayName = CSTRING(30Rnd_762x51_FAL_AP_Mag_Name);
+    ammo = "ACE_762x51_Ball_M993_AP";
+    lastRoundsTracer = 2;
+};
