+++ addons/ammunition/762x54r.hpp
@@ -1,0 +1,8 @@
+// 10Rnd 762x54 Magazines
+class CLASS(10Rnd_762x54_SVD_AP_Mag): 10Rnd_762x54_Mag {
+    author = "TyroneMF";
+    scope = 2;
+    displayName = CSTRING(10Rnd_762x54_SVD_AP_Mag_Name);
+    ammo = QCLASS(762x54_AP);
+    lastRoundsTracer = 1;
+};
